refactor: move _slack up a level
